requirements.txt
cachetools==5.3.1
aiohttp==3.9.0
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.146
botocore==1.29.146
bs4==0.0.1
faiss-cpu==1.7.4
gensim==4.3.2
hdbscan==0.8.33
jira==3.5.0
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonmerge==1.9.0
jsonschema==4.17.3
kombu==5.2.4
langchain==0.1.4
langchain-community==0.0.16
llama_cpp_python==0.2.15
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
Pillow==10.2.0
pydantic==1.10.8
PyJWT==2.4.0
PyPDF2==3.0.1
pytest==7.3.2
python-dotenv==1.0.0
redis==4.5.5
regex==2023.5.5
sentence-transformers==2.2.2
spacy==3.7.2
uvicorn==0.22.0
slack-sdk==3.26.1
pylint==2.17.4
pytest-cov==4.1.0
pytest-mock==3.11.1
tensorflow==2.14.0
transformers==4.36.0
torch==2.0.1
pypdf==3.17.0
asyncio==3.4.3
prometheus-client==0.17.1
rdflib==7.0.0
pyshacl==0.25.0
aiofiles==23.2.1
pytest-asyncio==0.23.2
google-cloud-storage==2.14.0
google-cloud==0.34.0
pymupdf==1.23.7
pydub==0.25.1
SpeechRecognition==3.10.1
pytesseract==0.3.10
pytextract==2.0.1
pandas==2.1.4
python-pptx==0.6.23
tika==2.6.0
openpyxl==3.1.2
coverage==7.3.3
azure-storage-blob==12.19.0
moviepy==1.0.3
python-docx==1.1.0
setuptools==69.0.2
psutil==5.9.7
dropbox==11.36.2
requests==2.31.0
google-api-python-client==2.105.0
rapidocr-onnxruntime==1.3.9
pybase64==1.3.1
pdfminer==20191125
requests_html==0.10.0
unidecode==1.3.7